+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [#3069](https://github.com/lapce/lapce/pull/3069): Allow color variables for themes
 - [#3086](https://github.com/lapce/lapce/pull/3086): Add folding to panels
 - [#3095](https://github.com/lapce/lapce/pull/3095): Add option to hide the Open Editors section in the explorer
+- [#3096](https://github.com/lapce/lapce/pull/3096): Add source control status coloring to file explorer
 
 ### Bug Fixes
 - [#2779](https://github.com/lapce/lapce/pull/2779): Fix files detection on fresh git/VCS repository